Separated server.js to multiple files
