+++ Vagrantfile
@@ -80,10 +80,8 @@ Vagrant.configure(2) do |config|
     echo "deb http://repo.mongodb.org/apt/ubuntu "$(lsb_release -sc)"/mongodb-org/3.0 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-org-3.0.list
     sudo apt-get update
     sudo apt-get install -y mongodb-org
-    #sudo apt-get install -y couchdb
 
-    #sudo sed -i 's/bind_address = 127.0.0.1/bind_address = 0.0.0.0/g' /etc/couchdb/default.ini
-    #sudo service couchdb restart
+    sudo npm install -g mocha
 
     echo node.js version:
     node -v
