+++ config/database.js
@@ -1,0 +1,5 @@
+// config/database.js
+
+module.exports = {
+    url : 'mongodb://localhost/grdmeUsers'
+};
