--- public/index.html
@@ -1,9 +1,0 @@
-<html>
-
-    <head>
-
-    </head>
-    <body>
-        <h1>HELLO, WORLD!</h1>
-    </body>
-</html>
